Found the testnet NewAddress bug :)
