+++ src/base58.h
@@ -249,204 +249,4 @@ public:
     bool operator> (const CBase58Data& b58) const { return CompareTo(b58) >  0; }
 };
 
-/** base58-encoded Primecoin addresses.
- * Public-key-hash-addresses have version 0 (or 111 testnet).
- * The data vector contains RIPEMD160(SHA256(pubkey)), where pubkey is the serialized public key.
- * Script-hash-addresses have version 5 (or 196 testnet).
- * The data vector contains RIPEMD160(SHA256(cscript)), where cscript is the serialized redemption script.
- */
-class PrimecoinAddress;
-class PrimecoinAddressVisitor : public boost::static_visitor<bool>
-{
-private:
-    PrimecoinAddress *addr;
-public:
-    PrimecoinAddressVisitor(PrimecoinAddress *addrIn) : addr(addrIn) { }
-    bool operator()(const CKeyID &id) const;
-    bool operator()(const CScriptID &id) const;
-    bool operator()(const CNoDestination &no) const;
-};
-
-class PrimecoinAddress : public CBase58Data
-{
-public:
-    enum
-    {
-        PUBKEY_ADDRESS = 23,  // primecoin: pubkey address starts with 'A'
-        SCRIPT_ADDRESS = 83,  // primecoin: script address starts with 'a'
-        PUBKEY_ADDRESS_TEST = 111,
-        SCRIPT_ADDRESS_TEST = 196,
-    };
-
-    bool Set(const CKeyID &id) {
-        SetData(fTestNet ? PUBKEY_ADDRESS_TEST : PUBKEY_ADDRESS, &id, 20);
-        return true;
-    }
-
-    bool Set(const CScriptID &id) {
-        SetData(fTestNet ? SCRIPT_ADDRESS_TEST : SCRIPT_ADDRESS, &id, 20);
-        return true;
-    }
-
-    bool Set(const CTxDestination &dest)
-    {
-        return boost::apply_visitor(PrimecoinAddressVisitor(this), dest);
-    }
-
-    bool IsValid() const
-    {
-        unsigned int nExpectedSize = 20;
-        bool fExpectTestNet = false;
-        switch(nVersion)
-        {
-            case PUBKEY_ADDRESS:
-                nExpectedSize = 20; // Hash of public key
-                fExpectTestNet = false;
-                break;
-            case SCRIPT_ADDRESS:
-                nExpectedSize = 20; // Hash of CScript
-                fExpectTestNet = false;
-                break;
-
-            case PUBKEY_ADDRESS_TEST:
-                nExpectedSize = 20;
-                fExpectTestNet = true;
-                break;
-            case SCRIPT_ADDRESS_TEST:
-                nExpectedSize = 20;
-                fExpectTestNet = true;
-                break;
-
-            default:
-                return false;
-        }
-        return fExpectTestNet == fTestNet && vchData.size() == nExpectedSize;
-    }
-
-    PrimecoinAddress()
-    {
-    }
-
-    PrimecoinAddress(const CTxDestination &dest)
-    {
-        Set(dest);
-    }
-
-    PrimecoinAddress(const std::string& strAddress)
-    {
-        SetString(strAddress);
-    }
-
-    PrimecoinAddress(const char* pszAddress)
-    {
-        SetString(pszAddress);
-    }
-
-    CTxDestination Get() const {
-        if (!IsValid())
-            return CNoDestination();
-        switch (nVersion) {
-        case PUBKEY_ADDRESS: {
-            uint160 id;
-            memcpy(&id, &vchData[0], 20);
-            return CKeyID(id);
-        }
-        case SCRIPT_ADDRESS: {
-            uint160 id;
-            memcpy(&id, &vchData[0], 20);
-            return CScriptID(id);
-        }
-        }
-        return CNoDestination();
-    }
-
-    bool GetKeyID(CKeyID &keyID) const {
-        if (!IsValid())
-            return false;
-        switch (nVersion) {
-        case PUBKEY_ADDRESS: {
-            uint160 id;
-            memcpy(&id, &vchData[0], 20);
-            keyID = CKeyID(id);
-            return true;
-        }
-        default: return false;
-        }
-    }
-
-    bool IsScript() const {
-        if (!IsValid())
-            return false;
-        switch (nVersion) {
-        case SCRIPT_ADDRESS: {
-            return true;
-        }
-        default: return false;
-        }
-    }
-};
-
-bool inline PrimecoinAddressVisitor::operator()(const CKeyID &id) const         { return addr->Set(id); }
-bool inline PrimecoinAddressVisitor::operator()(const CScriptID &id) const      { return addr->Set(id); }
-bool inline PrimecoinAddressVisitor::operator()(const CNoDestination &id) const { return false; }
-
-/** A base58-encoded secret key */
-class PrimecoinSecret : public CBase58Data
-{
-public:
-    void SetSecret(const CSecret& vchSecret, bool fCompressed)
-    {
-        assert(vchSecret.size() == 32);
-        SetData(128 + (fTestNet ? PrimecoinAddress::PUBKEY_ADDRESS_TEST : PrimecoinAddress::PUBKEY_ADDRESS), &vchSecret[0], vchSecret.size());
-        if (fCompressed)
-            vchData.push_back(1);
-    }
-
-    CSecret GetSecret(bool &fCompressedOut)
-    {
-        CSecret vchSecret;
-        vchSecret.resize(32);
-        memcpy(&vchSecret[0], &vchData[0], 32);
-        fCompressedOut = vchData.size() == 33;
-        return vchSecret;
-    }
-
-    bool IsValid() const
-    {
-        bool fExpectTestNet = false;
-        switch(nVersion)
-        {
-            case (128 + PrimecoinAddress::PUBKEY_ADDRESS):
-                break;
-
-            case (128 + PrimecoinAddress::PUBKEY_ADDRESS_TEST):
-                fExpectTestNet = true;
-                break;
-
-            default:
-                return false;
-        }
-        return fExpectTestNet == fTestNet && (vchData.size() == 32 || (vchData.size() == 33 && vchData[32] == 1));
-    }
-
-    bool SetString(const char* pszSecret)
-    {
-        return CBase58Data::SetString(pszSecret) && IsValid();
-    }
-
-    bool SetString(const std::string& strSecret)
-    {
-        return SetString(strSecret.c_str());
-    }
-
-    PrimecoinSecret(const CSecret& vchSecret, bool fCompressed)
-    {
-        SetSecret(vchSecret, fCompressed);
-    }
-
-    PrimecoinSecret()
-    {
-    }
-};
-
 #endif // __BASE58_H__
