+++ src/primecoin_address.cpp
@@ -1,0 +1,5 @@
+// Copyright (c) 2018 Chapman Shoop
+// See COPYING for license.
+
+#include "primecoin_address.h"
+
