+++ src/primecoin_secret.cpp
@@ -1,0 +1,6 @@
+// Copyright (c) 2018 Chapman Shoop
+// See COPYING for license.
+
+#include "primecoin_secret.h"
+
+
